Some new in databaseModule. Begin TreeManager.kt. Fix Int/Long troubles

+++ database/src/main/java/com/digitalsamurai/graphlib/database/room/nodes/node/entity/ChildNodes.kt
@@ -4,5 +4,5 @@ import com.google.gson.annotations.SerializedName
 
 data class ChildNodes(
     @SerializedName("childs")
-    var childs : ArrayList<Int>
+    var childs : ArrayList<Long>
 )
